refactor: query 및 타입 변환 함수 분리
